# api-documentation.yaml
openapi: 3.0.0
info:
  title: "VoxDem Chart API - Simplificada para Gráficos"
  version: "2.1.0"
  description: |
    API simplificada para geração de gráficos com dados da pesquisa VoxDem.
    Fornece dados otimizados para Chart.js com:
    - Lista de perguntas disponíveis
    - Lista de atributos de perfil demográfico
    - Dados para gráficos simples (sem cruzamento)
    - Dados para gráficos com cruzamento por perfil demográfico
    - Labels detalhadas com códigos de resposta
    - Ordenação automática por answer_code crescente
  contact:
    name: "VoxDem API Support"
    email: "suporte@voxdem.com"
servers:
  - url: "http://localhost:3000"
    description: "Servidor de Desenvolvimento"
...
paths:
...
  /api/chart/{questionCode}:
    get:
      summary: "Dados para gráfico simples (sem cruzamento)"
      description: |
        Retorna dados otimizados para Chart.js de uma pergunta específica.
        
        **Novidades v2.1.0:**
        - Labels detalhadas com códigos de resposta
        - Ordenação automática por answer_code crescente
        - Compatibilidade mantida com Chart.js
      tags: 
        - "Charts"
      parameters:
        - name: questionCode
          in: path
          required: true
          description: "Código da pergunta (ex: P01, P23A)"
          schema:
            type: string
            pattern: "^P[0-9]+[A-Z]*$"
          example: "P01"
      responses:
        '200':
          description: "Dados do gráfico retornados com sucesso"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ChartData'
              example:
                success: true
                data:
                  question:
                    code: "P01"
                    text: "De uma maneira geral, o(a) Sr.(a) está muito satisfeito(a)..."
                  chartData:
                    labels: ["Muito satisfeito", "Satisfeito", "Insatisfeito", "Muito insatisfeito"]
                    labelsDetailed:
                      - code: "1"
                        label: "Muito satisfeito"
                        display: "1 - Muito satisfeito"
                      - code: "2"
                        label: "Satisfeito"
                        display: "2 - Satisfeito"
                      - code: "3"
                        label: "Insatisfeito"
                        display: "3 - Insatisfeito"
                      - code: "4"
                        label: "Muito insatisfeito"
                        display: "4 - Muito insatisfeito"
                    datasets:
                      - data: [89, 567, 642, 265]
                        backgroundColor: ["#FF6384", "#36A2EB", "#FFCE56", "#4BC0C0"]
                        borderColor: ["#FF6384", "#36A2EB", "#FFCE56", "#4BC0C0"]
                        borderWidth: 1
                  totalResponses: 1563
                generatedAt: "2025-09-08T10:30:00.000Z"
        '400':
          description: "Código de pergunta inválido"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Pergunta não encontrada"
  /api/chart/{questionCode}/{profileAttribute}:
    get:
      summary: "Dados para gráfico com cruzamento por perfil"
      description: |
        Retorna dados para gráfico de barras agrupadas por perfil demográfico.
        
        **Novidades v2.1.0:**
        - Labels detalhadas com códigos de resposta
        - Ordenação automática por answer_code crescente
        - Mantém compatibilidade total com Chart.js
      tags: 
        - "Charts"
      parameters:
        - name: questionCode
          in: path
          required: true
          description: "Código da pergunta (ex: P01, P23A)"
          schema:
            type: string
            pattern: "^P[0-9]+[A-Z]*$"
          example: "P01"
        - name: profileAttribute
          in: path
          required: true
          description: "Atributo de perfil para cruzamento"
          schema:
            type: string
            enum: 
              - gender
              - age_range
              - education
              - race
              - region
              - state
          example: "gender"
      responses:
        '200':
          description: "Dados do gráfico com perfil retornados com sucesso"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProfileChartData'
              example:
                success: true
                data:
                  question:
                    code: "P01"
                    text: "De uma maneira geral, o(a) Sr.(a) está muito satisfeito(a)..."
                  profileAttribute: "gender"
                  chartData:
                    labels: ["Muito satisfeito", "Satisfeito", "Insatisfeito", "Muito insatisfeito"]
                    labelsDetailed:
                      - code: "1"
                        label: "Muito satisfeito"
                        display: "1 - Muito satisfeito"
                      - code: "2"
                        label: "Satisfeito"
                        display: "2 - Satisfeito"
                      - code: "3"
                        label: "Insatisfeito"
                        display: "3 - Insatisfeito"
                      - code: "4"
                        label: "Muito insatisfeito"
                        display: "4 - Muito insatisfeito"
                    datasets:
                      - label: "Masculino"
                        data: [45, 280, 320, 135]
                        backgroundColor: "#FF6384"
                        borderColor: "#FF6384"
                        borderWidth: 1
                      - label: "Feminino"
                        data: [44, 287, 322, 130]
                        backgroundColor: "#36A2EB"
                        borderColor: "#36A2EB"
                        borderWidth: 1
                  totalResponses: 1563
                generatedAt: "2025-09-08T10:30:00.000Z"
        '400':
          description: "Parâmetros inválidos"
        '404':
          description: "Pergunta ou atributo não encontrado"
components:
  schemas:
...
    ChartData:
      type: object
      description: "Dados otimizados para Chart.js"
      properties:
        question:
          type: object
          properties:
            code:
              type: string
            text:
              type: string
            category:
              type: string
              nullable: true
        chartData:
          type: object
          description: "Dados otimizados para Chart.js com labels detalhadas"
          properties:
            labels:
              type: array
              items:
                type: string
              description: "Labels simples para compatibilidade com Chart.js"
            labelsDetailed:
              type: array
              items:
                type: object
                properties:
                  code:
                    type: string
                    description: "Código da resposta (answer_code)"
                    example: "1"
                  label:
                    type: string
                    description: "Texto da resposta"
                    example: "Muito satisfeito"
                  display:
                    type: string
                    description: "Formato combinado: código - texto"
                    example: "1 - Muito satisfeito"
              description: "Labels detalhadas com códigos ordenadas por answer_code crescente"
            datasets:
              type: array
              items:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: integer
                  backgroundColor:
                    type: array
                    items:
                      type: string
                  borderColor:
                    type: array
                    items:
                      type: string
                  borderWidth:
                    type: integer
        totalResponses:
          type: integer
    ProfileChartData:
      type: object
      description: "Dados de gráfico com perfil demográfico"
      properties:
        question:
          type: object
        profileAttribute:
          type: string
        chartData:
          type: object
          description: "Dados para gráfico de barras agrupadas com labels detalhadas"
          properties:
            labels:
              type: array
              items:
                type: string
              description: "Labels simples para compatibilidade com Chart.js"
            labelsDetailed:
              type: array
              items:
                type: object
                properties:
                  code:
                    type: string
                    description: "Código da resposta (answer_code)"
                  label:
                    type: string
                    description: "Texto da resposta"
                  display:
                    type: string
                    description: "Formato combinado: código - texto"
              description: "Labels detalhadas com códigos ordenadas por answer_code crescente"
            datasets:
              type: array
              items:
                type: object
                properties:
                  label:
                    type: string
                  data:
                    type: array
                    items:
                      type: integer
                  backgroundColor:
                    type: string
                  borderColor:
                    type: string
        totalResponses:
          type: integer
    ApiResponse:
      type: object
      description: "Resposta padrão da API"
      properties:
        success:
          type: boolean
          example: true
        data:
          description: "Dados retornados"
        count:
          type: integer
          description: "Número de itens"
        generatedAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      description: "Resposta de erro"
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        message:
          type: string